+++ src/test/main/Main.test.tsx
@@ -2,7 +2,7 @@ import { Main } from '@pages/main';
 import { act, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { BrowserRouter } from 'react-router-dom';
-import { server } from 'src/mocks/browsers';
+import { server } from 'src/mocks/server';
 import { HttpResponse, http } from 'msw';
 import staticsData from 'public/data/main/staticsData.json';
